socialrcb_backend/requirements.txt
APScheduler==3.9.1
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
fonttools==4.31.2
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
mysql-connector==2.2.9
mysql-connector-python==8.0.29
mysqlclient==2.1.0
numpy==1.22.3
openapi-codec==1.3.2
packaging==21.3
Pillow==9.1.0
psycopg2==2.9.3
pycodestyle==2.8.0
PyJWT==2.3.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.9
